+++ src/built-ins/index.js
@@ -1,6 +1,8 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
+export * from '#x/AccessLogToFile';
+export * from '#x/AccessLogToSyslog';
 export * from '#x/EventFan';
 export * from '#x/HostRouter';
 export * from '#x/MemoryMonitor';
@@ -9,8 +11,6 @@ export * from '#x/ProcessIdFile';
 export * from '#x/ProcessInfoFile';
 export * from '#x/RateLimiter';
 export * from '#x/Redirector';
-export * from '#x/RequestLogger';
-export * from '#x/RequestSyslogger';
 export * from '#x/SerialRouter';
 export * from '#x/SimpleResponse';
 export * from '#x/StaticFiles';