+++ src/sys-framework/export/NetworkEndpoint.js
@@ -4,8 +4,7 @@
 import { TreePathKey } from '@this/collections';
 import { BaseComponent, BaseNamedConfig, Names } from '@this/compote';
 import { FormatUtils } from '@this/loggy-intf';
-import { IntfRateLimiter, IntfRequestLogger, ProtocolWrangler,
-  ProtocolWranglers }
+import { IntfAccessLog, IntfRateLimiter, ProtocolWrangler, ProtocolWranglers }
   from '@this/net-protocol';
 import { DispatchInfo, HostUtil, IntfRequestHandler, OutgoingResponse, UriUtil }
   from '@this/net-util';
@@ -97,13 +96,13 @@ export class NetworkEndpoint extends BaseComponent {
       interface: iface,
       protocol,
       services: {
-        rateLimiter:   rateLimiterName = null,
-        requestLogger: requestLoggerName = null
+        accessLog:   accessLogName = null,
+        rateLimiter: rateLimiterName = null
       }
     } = this.config;
 
-    const rateLimiter   = context.getComponentOrNull(rateLimiterName,   BaseService, IntfRateLimiter);
-    const requestLogger = context.getComponentOrNull(requestLoggerName, BaseService, IntfRequestLogger);
+    const rateLimiter = context.getComponentOrNull(rateLimiterName, BaseService, IntfRateLimiter);
+    const accessLog   = context.getComponentOrNull(accessLogName,   BaseService, IntfAccessLog);
 
     const hmOpt = {};
     if (this.config.requiresCertificates()) {
@@ -112,9 +111,9 @@ export class NetworkEndpoint extends BaseComponent {
     }
 
     const wranglerOptions = {
+      accessLog,
       rateLimiter,
       requestHandler: this,
-      requestLogger,
       protocol,
       interface: iface,
       ...hmOpt